requirements.txt
torch
torchvision
Pillow
numpy
pandas
sklearn
flask